ControllerReport廃止。Entrypointをもとに出力。

+++ jig-core/src/main/java/org/dddjava/jig/domain/model/parts/classes/method/DecisionNumber.java
@@ -18,7 +18,7 @@ public class DecisionNumber {
         return value > 0;
     }
 
-    public double intValue() {
+    public int intValue() {
         return value;
     }
 }
